* Removed package `dsn` * Added `sugar.DSN` helper for make dsn (connection string)
